move auth into separate root directory because I want to automate the deployment of the infrastructure by calling terraform apply within github actions.
As the github repo will only have access to a service account contained in the wif_pool, all of the variables and permissions required will need to first be initialised outside of github. Then, the rest of the terraform depolyment can proceed solely on github. Therefore need to separate the main terraform config from that auth init config.